requirements.txt
browser-use==0.1.37
pyperclip==1.9.0
gradio==5.10.0
json-repair
langchain-mistralai==0.2.4
google-generativeai>=0.3.0
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.22.0
